% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmsLogin.R
\name{cdmsLogin}
\alias{cdmsLogin}
\title{cdmsLogin: Login to CDMS.}
\usage{
cdmsLogin(username, api_key, cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{username}{for CDMS.}

\item{api_key}{provided through the user preferences page of the CDMS user-interface.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
The function passes username and password creditianls to the
  specified Centralized Data Management System (CDMS). After being
  successfull logged in a message will display in the console stating the
  full name of the user and a cookie will be stored for all future api calls
  for data retrieval. The cookie expires in 24-hours and requires
  \code{cdmsLogin()} to be re-run to gain access.
}
\examples{
cdmsLogin(username = 'username', api_key = 'key',
 cdms_host = 'https://cdms.nptfisheries.org')

}
\author{
Ryan Kinzer
}
