% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Activities.R
\name{get_Activities}
\alias{get_Activities}
\title{get_Activities:}
\usage{
get_Activities(datasetID, cdms_host = "https://npt-cdms.nezperce.org")
}
\arguments{
\item{datasetID}{for CDMS dataset.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get all activities in CDMS.
}
\author{
Ryan Kinzer
}
