% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AllFiles.r
\name{get_AllFiles}
\alias{get_AllFiles}
\title{get_AllFiles:}
\usage{
get_AllFiles(cdms_host = "https://npt-cdms.nezperce.org/")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
gets all files stored in CDMS.
}
\author{
Ryan Kinzer, Tyler Stright
}
