% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CarcassData.R
\name{get_CarcassData}
\alias{get_CarcassData}
\title{get_CarcassData:}
\usage{
get_CarcassData(
  SurveyYear = NULL,
  Project = c("All", "IRSSM", "JCAPE", "NPTH"),
  LocationLabel = NULL,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{Project}{desired project acronym.  NULL returns all.}

\item{LocationLabel}{desired Location label. NULL returns all.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}

\item{SurveYear}{desired survey year (YYYY).  NULL returns all.}
}
\description{
get data from
}
\author{
Tyler Stright
}
