% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CarcassData_NEOR.R
\name{get_CarcassData_NEOR}
\alias{get_CarcassData_NEOR}
\title{get_CarcassData_NEOR:}
\usage{
get_CarcassData_NEOR(
  SurveyYear = NULL,
  GRSME_ONLY = TRUE,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{GRSME_ONLY}{TRUE: Return only GRSME data. FALSE: Return all ODFW data.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}

\item{SurveYear}{desired survey year (YYYY).  NULL returns all.}
}
\description{
Retrieve North East Oregon spawning ground survey carcass data from CDMS.
}
\author{
Tyler Stright
}
