% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DABOMrecruits.R
\name{get_DABOMrecruits}
\alias{get_DABOMrecruits}
\title{get_DABOMrecruits:}
\usage{
get_DABOMrecruits(
  Variable = c("All", "Recruits", "Spawners", "lambda"),
  BroodYear = NULL,
  Species = c("All", "Chinook salmon", "Steelhead"),
  Run = c("All", "Spring/Summer", "Summer"),
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{Variable}{filter to return only a specific variable. NULL returns all variables.}

\item{BroodYear}{four digit year filter (YYYY) on Brood Year. NULL returns all years.}

\item{Species}{filter to return data from a single species. NULL returns all species.}

\item{Run}{filter to return data from a single run of fish. NULL returns all runs.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
Retrieve DABOM (Dam Adult Branch Occupancy Model) recruits estimates from CDMS.
}
\author{
Tyler Stright
}
