% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DatasetView.R
\name{get_DatasetView}
\alias{get_DatasetView}
\title{get_DatasetView:}
\usage{
get_DatasetView(
  datastoreID,
  projectID = NULL,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{datastoreID}{for CDMS dataset. see get_Datastores()}

\item{projectID}{for project Id of interest in CDMS project table. see get_Projects()}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get all data for a given dataset (the datastore), filtered by project if desired.
}
\author{
Ryan Kinzer
}
