% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Datastores.R
\name{get_Datastores}
\alias{get_Datastores}
\title{get_Datastores:}
\usage{
get_Datastores(cdms_host = "https://npt-cdms.nezperce.org")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
gets all the available datasets from the datastore table.
}
\author{
Ryan Kinzer
}
