% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DetailRecords.R
\name{get_DetailRecords}
\alias{get_DetailRecords}
\title{get_DetailRecords:}
\usage{
get_DetailRecords(activityID, cdms_host = "https://npt-cdms.nezperce.org")
}
\arguments{
\item{activityID}{for CDMS dataset.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get detail records for a given activity.
}
\author{
Ryan Kinzer
}
