% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FallRR.R
\name{get_FallRR}
\alias{get_FallRR}
\title{get_FallRR:}
\usage{
get_FallRR(
  BroodYear = NULL,
  ReturnYear = NULL,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{BroodYear}{four digit year filter (YYYY) on Brood Year. NULL returns all years.}

\item{ReturnYear}{four digit year filter (YYYY) on Return Year. NULL returns all years.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
Retrieve Fall Chinook Run Reconstruction data from CDMS.
}
\author{
Tyler Stright
}
