% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FINSvalues.R
\name{get_FINSvalues}
\alias{get_FINSvalues}
\title{get_FINSvalues:}
\usage{
get_FINSvalues(
  Module = c("Trapping", "Spawning"),
  Field = c("Facility", "Spawning Location", "Species", "Origin", "Stock"),
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{Module}{desired FINS module: Trapping or Spawning.}

\item{Field}{desired field chosen to display possible values.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
helper function to provide acceptable values for FINS data queries:
get_WeirData() and get_SpawningData().
}
\author{
Tyler Stright
}
