% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HeaderRecords.R
\name{get_HeaderRecords}
\alias{get_HeaderRecords}
\title{get_HeaderRecords:}
\usage{
get_HeaderRecords(datasetID, cdms_host = "https://npt-cdms.nezperce.org")
}
\arguments{
\item{datasetID}{for CDMS dataset.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get header records for a given activity.
}
\author{
Ryan Kinzer
}
