% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_JuvSurvival.R
\name{get_JuvSurvival}
\alias{get_JuvSurvival}
\title{get_JuvSurvival:}
\usage{
get_JuvSurvival(
  RST = c("All", "Imnaha River", "Johnson Creek", "Lolo Creek", "Newsome Creek",
    "Secesh River", "South Fork Clearwater River"),
  SpeciesRun = NULL,
  MigratoryYear = NULL,
  BroodYear = NULL,
  Origin = c("All", "Hatchery", "Natural"),
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get data from
}
\author{
Tyler Stright
}
