% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LocationTypes.R
\name{get_LocationTypes}
\alias{get_LocationTypes}
\title{get_LocationTypes:}
\usage{
get_LocationTypes(cdms_host = "https://npt-cdms.nezperce.org")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get all available location types from LocationType table.
}
\author{
Ryan Kinzer
}
