% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ProjectData.R
\name{get_ProjectData}
\alias{get_ProjectData}
\title{get_ProjectData:}
\usage{
get_ProjectData(datasetID, cdms_host = "https://npt-cdms.nezperce.org")
}
\arguments{
\item{datasetID}{for CDMS dataset.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get all records for a given project dataset.
}
\author{
Ryan Kinzer
}
