% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ProjectDatasets.R
\name{get_ProjectDatasets}
\alias{get_ProjectDatasets}
\title{get_ProjectDatasets:}
\usage{
get_ProjectDatasets(projectID, cdms_host = "https://npt-cdms.nezperce.org")
}
\arguments{
\item{projectID}{for CDMS projects.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get a list of datasets associated with a project.
}
\author{
Ryan Kinzer
}
