% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ProjectFiles.r
\name{get_ProjectFiles}
\alias{get_ProjectFiles}
\title{get_ProjectFiles:}
\usage{
get_ProjectFiles(ProjectId, cdms_host = "https://npt-cdms.nezperce.org/")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
gets all files associated with a project.
}
\author{
Ryan Kinzer, Tyler Stright
}
