% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Projects.R
\name{get_Projects}
\alias{get_Projects}
\title{get_Projects:}
\usage{
get_Projects(cdms_host = "https://npt-cdms.nezperce.org")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get all the available projects in CDMS.
}
\author{
Ryan Kinzer
}
