% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ReddData.R
\name{get_ReddData}
\alias{get_ReddData}
\title{get_ReddData:}
\usage{
get_ReddData(
  SurveyYear = NULL,
  Project = c("All", "IRSSSM", "JCAPE", "NPTH"),
  LocationLabel = NULL,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{SurveyYear}{desired survey year (YYYY).  NULL returns all.}

\item{Project}{desired project acronym.  NULL returns all.}

\item{LocationLabel}{desired Location label. NULL returns all.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
Retrieve spawning ground survey redd data from CDMS (does not include NE Oregon).
}
\author{
Tyler Stright
}
