% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Users.r
\name{get_Users}
\alias{get_Users}
\title{get_Users:}
\usage{
get_Users(cdms_host = "https://npt-cdms.nezperce.org/")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
gets information for users with CDMS login credentials.
}
\author{
Ryan Kinzer, Tyler Stright
}
