% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Waterbodies.R
\name{get_Waterbodies}
\alias{get_Waterbodies}
\title{get_Waterbodies:}
\usage{
get_Waterbodies(cdms_host = "https://npt-cdms.nezperce.org")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get all CDMS water bodies.
}
\author{
Ryan Kinzer
}
