% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_WaterTempData.R
\name{get_WaterTempData}
\alias{get_WaterTempData}
\title{get_WaterTempData()}
\usage{
get_WaterTempData(
  date_begin,
  date_end,
  locationID = NULL,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{date_begin}{First date of desired date range.}

\item{date_end}{Final Date of desired date range.}

\item{locationID}{from location table.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
Retrieve CDMS water temperature data between two dates.
}
\author{
Tyler Stright
}
