% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivities.R
\name{getActivities}
\alias{getActivities}
\title{getActivities:}
\usage{
getActivities(datasetID, cdms_host = "https://npt-cdms.nezperce.org")
}
\arguments{
\item{datasetID}{for CDMS dataset.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get all activities in CDMS.
}
\author{
Ryan Kinzer
}
