% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllFiles.r
\name{getAllFiles}
\alias{getAllFiles}
\title{getAllFiles:}
\usage{
getAllFiles(cdms_host = "https://npt-cdms.nezperce.org/")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
gets all files stored in CDMS.
}
\author{
Ryan Kinzer, Tyler Stright
}
