% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasetView.R
\name{getDatasetView}
\alias{getDatasetView}
\title{getDatasetView:}
\usage{
getDatasetView(datastoreID, projectID = NULL, waterbodyID = NULL,
  locationID = NULL, cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{datastoreID}{for CDMS dataset.}

\item{projectID}{}

\item{waterbodyID}{}

\item{locationID}{}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{

}
\examples{
getDatasetView(datastoreID, cdms_host = 'https://cdms.nptfisheries.org')

}
\author{
Ryan Kinzer
}
