% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDetailRecords.R
\name{getDetailRecords}
\alias{getDetailRecords}
\title{getDetailRecords:}
\usage{
getDetailRecords(activityID, cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{activityID}{for CDMS dataset.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get detail records for a given activity.
}
\author{
Ryan Kinzer
}
