% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFCRRdata.R
\name{getFCRRdata}
\alias{getFCRRdata}
\title{get}
\usage{
getFCRRdata(
  BroodYear = NULL,
  ReturnYear = NULL,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{BroodYear}{four digit year filter (YYYY) on Brood Year. NULL returns all years.}

\item{ReturnYear}{four digit year filter (YYYY) on Return Year. NULL returns all years.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get data from
}
\author{
Tyler Stright
}
