% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFINSspawningData.R
\name{getFINSspawningData}
\alias{getFINSspawningData}
\title{get}
\usage{
getFINSspawningData(
  SpawnLocation = NULL,
  Stock = NULL,
  Species = NULL,
  Run = c("All", NA, "Spring", "Summer", "Fall", "Winter"),
  Sex = c("All", "Female", "Male", "Unknown"),
  Origin = NULL,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{SpawnLocation}{desired Spawning Location. NULL returns all. Discover values: getFINSvalues()}

\item{Species}{desired species. NULL returns all. Discover values: getFINSvalues()}

\item{Run}{desired run of fish. NULL returns all.}

\item{Sex}{desired sex of fish. NULL returns all.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get data from
}
\author{
Tyler Stright
}
