% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFINSvalues.R
\name{getFINSvalues}
\alias{getFINSvalues}
\title{getFINSvalues}
\usage{
getFINSvalues(
  module = c("Trapping", "Spawning"),
  field = c("Facility", "Spawning Location", "Species", "Origin", "Stock"),
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{module}{desired FINS module: Trapping or Spawning.}

\item{field}{desired field chosen to display possible values.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
helper function to provide acceptable values for FINS data queries:
getFINSweirData() and getFINSspawningData().
}
\author{
Tyler Stright
}
