% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeaderRecords.R
\name{getHeaderRecords}
\alias{getHeaderRecords}
\title{getHeaderRecords:}
\usage{
getHeaderRecords(datasetID, cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{datasetID}{for CDMS dataset.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{

}
\examples{
getHeaderRecords(datasetID, cdms_host = 'https://cdms.nptfisheries.org')

}
\author{
Ryan Kinzer
}
