% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIPTDSescData.R
\name{getIPTDSescData}
\alias{getIPTDSescData}
\title{get}
\usage{
getIPTDSescData(
  Variable = c("All", "Female Escapement", "Age Escapement", "Female Proportion",
    "Age Proportion", "Population Escapement", "Site Escapement", "Detection Efficiency"),
  SpawnYear = NULL,
  Species = c("All", "Chinook salmon", "Steelhead"),
  Run = c("All", "Spring/Summer", "Summer"),
  TRT_POPID = NULL,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{Variable}{filter to return only a specific variable. NULL returns all variables.}

\item{SpawnYear}{four digit year filter (YYYY) on Spawn Year. NULL returns all years.}

\item{Species}{filter to return data from a single species. NULL returns all species.}

\item{Run}{filter to return data from a single run of fish. NULL returns all runs.}

\item{TRT_POPID}{filter to return data from a single ICTRT population. NULL returns all populations.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get data from
}
\author{
Tyler Stright
}
