% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIPTDSlgrData.R
\name{getIPTDSlgrData}
\alias{getIPTDSlgrData}
\title{get}
\usage{
getIPTDSlgrData(
  SpawnYear = NULL,
  Species = c("All", "Chinook salmon", "Steelhead"),
  Run = c("All", "Spring/Summer", "Summer"),
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{SpawnYear}{four digit year filter (YYYY) on Brood Year. NULL returns all years.}

\item{Species}{filter to return data from a single species. NULL returns all species.}

\item{Run}{filter to return data from a single run of fish. NULL returns all runs.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get data from
}
\author{
Tyler Stright
}
