% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocations.R
\name{getLocations}
\alias{getLocations}
\title{getLocations:}
\usage{
getLocations(cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{

}
\examples{
getLocations(cdms_host = 'https://cdms.nptfisheries.org')

}
\author{
Ryan Kinzer
}
