% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocationTypes.R
\name{getLocationTypes}
\alias{getLocationTypes}
\title{getLocationTypes:}
\usage{
getLocationTypes(cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get all available location types from LocationType table.
}
\author{
Ryan Kinzer
}
