% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getP4data.R
\name{getP4data}
\alias{getP4data}
\title{getP4data()}
\usage{
getP4data(
  MRRProject = c("All", "CDR", "JLV", "IMN", "NPC", "SCS"),
  EventSite = c("All", "IMNTRP", "JOHTRP", "LOLTRP", "SECTRP", "SFCTRP"),
  EventType = c("All", "Recapture", "Mark", "Recovery", "Tally", "Passive Recapture"),
  SRRcode = NULL,
  MigrationYear = NULL,
  BroodYear = NULL,
  CalendarYear = NULL,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}

\item{migration_year}{Desired migration year (YYYY)}

\item{event_site}{Desired event site. Leave NULL to retrieve all sites.}
}
\description{
Retrieve p4 data (from CDMS) for all sites or a single site.
}
\author{
Tyler Stright
}
