% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProject.R
\name{getProject}
\alias{getProject}
\title{getProject:}
\usage{
getProject(projectID, cdms_host = "https://npt-cdms.nezperce.org")
}
\arguments{
\item{projectID}{for CDMS Project.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get all information associated with a given project
}
\author{
Ryan Kinzer, Tyler Stright
}
