% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectData.R
\name{getProjectData}
\alias{getProjectData}
\title{getProjectData:}
\usage{
getProjectData(datasetID, cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{datasetID}{for CDMS dataset.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get all records for a given project dataset.
}
\author{
Ryan Kinzer
}
