% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectFiles.r
\name{getProjectFiles}
\alias{getProjectFiles}
\title{getProjectFiles:}
\usage{
getProjectFiles(ProjectId, cdms_host = "https://npt-cdms.nezperce.org/")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
gets all files associated with a project.
}
\author{
Ryan Kinzer, Tyler Stright
}
