% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjects.R
\name{getProjects}
\alias{getProjects}
\title{getProjects:}
\usage{
getProjects(cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}

\item{datastoreID}{for CDMS dataset.}
}
\description{

}
\examples{
getProjects(cdms_host = 'https://cdms.nptfisheries.org')

}
\author{
Ryan Kinzer
}
