% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSGScarcassDataNEOR.R
\name{getSGScarcassDataNEOR}
\alias{getSGScarcassDataNEOR}
\title{get}
\usage{
getSGScarcassDataNEOR(
  SurveyYear = NULL,
  GRSME_ONLY = TRUE,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{GRSME_ONLY}{TRUE: Return only GRSME data. FALSE: Return all ODFW data.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}

\item{SurveYear}{desired survey year (YYYY).  NULL returns all.}
}
\description{
get data from
}
\author{
Tyler Stright
}
