% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsers.r
\name{getUsers}
\alias{getUsers}
\title{getUsers:}
\usage{
getUsers(cdms_host = "https://npt-cdms.nezperce.org/")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
gets information for users with CDMS login credentials.
}
\author{
Ryan Kinzer, Tyler Stright
}
