% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWaterbodies.R
\name{getWaterbodies}
\alias{getWaterbodies}
\title{getWaterbodies:}
\usage{
getWaterbodies(cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get all waterbodies listed in waterbodies table.
}
\author{
Ryan Kinzer
}
