% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ProjectLocations.R
\name{get_ProjectLocations}
\alias{get_ProjectLocations}
\title{get_ProjectLocations:}
\usage{
get_ProjectLocations(ProjectID, cdms_host = "https://npt-cdms.nezperce.org")
}
\arguments{
\item{ProjectID}{CDMS Project ID. see cdmsR::get_Projects()}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}

\item{DatastoreID}{CDMS Datastore ID. see cdmsR::get_Datastores().  NULL returns all.}
}
\description{
get Location information from CDMS for a specific project..
}
\author{
Tyler Stright
}
