% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_WaterTempData.R
\name{get_WaterTempData}
\alias{get_WaterTempData}
\title{get_WaterTempData:}
\usage{
get_WaterTempData(
  year,
  locationID = NULL,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{year}{Desired year of data (YYYY).}

\item{locationID}{Location ID of the desired location. See get_ProjectLocations()}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
Retrieve CDMS water temperature data from CDMS for a calendar year.
}
\author{
Tyler Stright
}
