% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasetView.R
\name{getDatasetView}
\alias{getDatasetView}
\title{getDatasetView:}
\usage{
getDatasetView(datastoreID, projectID = NULL, waterbodyID = NULL,
  locationID = NULL, cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{datastoreID}{for CDMS dataset.}

\item{projectID}{for project of interest in CDMS project table.}

\item{waterbodyID}{for stream in CDMS waterbodies table.}

\item{locationID}{from location table.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get the big bucket for a given dataset, filtered by project,
 waterbody and/or location.
}
\author{
Ryan Kinzer
}
