% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatastores.R
\name{getDatastores}
\alias{getDatastores}
\title{getDatastores:}
\usage{
getDatastores(cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
gets all the available datasets from the datastore table.
}
\author{
Ryan Kinzer
}
