% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getP4data.R
\name{getP4data}
\alias{getP4data}
\title{getP4data()}
\usage{
getP4data(
  MRRProject = c("All", "CDR", "JLV", "IMN", "NPC", "SCS"),
  EventSite = c("All", "JOHTRP", "SECTRP", "LSFTRP", "JOHNSC", "MCCA\\", \\"IMNTRP",
    "NPTH\\", \\"SFCTRP", "NEWSOC", "CLWRSF", "LOLOC", "LOLTRP"),
  EventType = c("All", "Recapture", "Mark", "Recovery", "Tally", "Passive Recapture"),
  CaptureMethod = c("All", "SCREWT", "FYKNET", "BSEINE", "DIPNET", "SHOCK"),
  SRRcode = NULL,
  MigrationYear = NULL,
  BroodYear = NULL,
  CalendarYear = NULL,
  cdms_host = "https://npt-cdms.nezperce.org"
)
}
\arguments{
\item{MRRProject}{PTAGIS Coordinator Code. NULL returns all.}

\item{EventSite}{Desired event site. NULL returns all sites.}

\item{EventType}{Desired event type. NULL returns all event types.}

\item{CaptureMethod}{Desired capture method. NULL returns all methods.}

\item{SRRcode}{Desired species/run/rear code. NULL returns all species.}

\item{MigrationYear}{Desired migration year (YYYY) NULL returns all years}

\item{BroodYear}{Desired brood Year (YYYY). NULL returns all years}

\item{CalendarYear}{Desired calendar year (YYYY) NULL returns all years.

#' @param cdms_host the web URL for the targeted CDMS user-interface page.}
}
\description{
Retrieve p4 data (from CDMS) for all sites or a single site.
}
\author{
Tyler Stright
}
