% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectData.R
\name{getProjectData}
\alias{getProjectData}
\title{getProjectData:}
\usage{
getProjectData(datasetID, cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}

\item{datastoreID}{for CDMS dataset.}
}
\description{

}
\examples{
getProjectData(datasetID, cdms_host = 'https://cdms.nptfisheries.org')

}
\author{
Ryan Kinzer
}
