% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectDatasets.R
\name{getProjectDatasets}
\alias{getProjectDatasets}
\title{getProjectDatasets:}
\usage{
getProjectDatasets(projectID,
  cdms_host = "https://cdms.nptfisheries.org")
}
\arguments{
\item{projectID}{for CDMS projects.}

\item{cdms_host}{the web URL for the targeted CDMS user-interface page.}
}
\description{
get a list of datasets associated with a project.
}
\author{
Ryan Kinzer
}
